// client-vms/src/gen-proto/nillion/meta/v1/tx_pb.ts
// @generated by protoc-gen-es v2.2.2 with parameter "target=ts"
// @generated from file nillion/meta/v1/tx.proto (package nillion.meta.v1, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv1";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file nillion/meta/v1/tx.proto.
 */
export const file_nillion_meta_v1_tx: GenFile = /*@__PURE__*/
  fileDesc("ChhuaWxsaW9uL21ldGEvdjEvdHgucHJvdG8SD25pbGxpb24ubWV0YS52MSJcCglNc2dQYXlGb3ISEAoIcmVzb3VyY2UYASABKAwSFAoMZnJvbV9hZGRyZXNzGAIgASgJEicKBmFtb3VudBgDIAMoCzIXLm5pbGxpb24ubWV0YS52MS5BbW91bnQiJwoGQW1vdW50Eg0KBWRlbm9tGAEgASgJEg4KBmFtb3VudBgCIAEoCUJ8ChNjb20ubmlsbGlvbi5tZXRhLnYxQgdUeFByb3RvUAGiAgNOTViqAg9OaWxsaW9uLk1ldGEuVjHKAg9OaWxsaW9uXE1ldGFcVjHiAhtOaWxsaW9uXE1ldGFcVjFcR1BCTWV0YWRhdGHqAhFOaWxsaW9uOjpNZXRhOjpWMWIGcHJvdG8z");

/**
 * @generated from message nillion.meta.v1.MsgPayFor
 */
export type MsgPayFor = Message<"nillion.meta.v1.MsgPayFor"> & {
  /**
   * @generated from field: bytes resource = 1;
   */
  resource: Uint8Array;

  /**
   * @generated from field: string from_address = 2;
   */
  fromAddress: string;

  /**
   * @generated from field: repeated nillion.meta.v1.Amount amount = 3;
   */
  amount: Amount[];
};

/**
 * Describes the message nillion.meta.v1.MsgPayFor.
 * Use `create(MsgPayForSchema)` to create a new message.
 */
export const MsgPayForSchema: GenMessage<MsgPayFor> = /*@__PURE__*/
  messageDesc(file_nillion_meta_v1_tx, 0);

/**
 * @generated from message nillion.meta.v1.Amount
 */
export type Amount = Message<"nillion.meta.v1.Amount"> & {
  /**
   * @generated from field: string denom = 1;
   */
  denom: string;

  /**
   * @generated from field: string amount = 2;
   */
  amount: string;
};

/**
 * Describes the message nillion.meta.v1.Amount.
 * Use `create(AmountSchema)` to create a new message.
 */
export const AmountSchema: GenMessage<Amount> = /*@__PURE__*/
  messageDesc(file_nillion_meta_v1_tx, 1);